Allow agentbeat to be importable.
